Don't lock down to a specific version of github-pages
Previously the tests were failing every time there was a new version of
the github-pages gem released because we were locking down to a specific
version and bundler was complaining.

This change means we don't automatically use the latest version from
GitHub, which means we will need to run `bundle update` a bit more
frequently to ensure we're running against the same version as Github.