Don't boot Jekyll if there are Gemfile.lock changes
We track the latest version of the `github-pages` gem so that we can be
sure that we're running against the same version as GitHub are using.
However this sometimes causes the tests to fail if you've updated
github-pages locally but not committed the resulting `Gemfile.lock`.

This change will refuse to boot up jekyll if it notices that you have
uncommitted changes to `Gemfile.lock`.