Add timeout for external link checker
This adds a 30 second timeout to html-proofer's external link checker.
Previously we had issues where a site was just hanging without returning
a response. This meant that the tests would never finish locally and
they'd timeout and fail on Travis.

By adding a 30 second timeout we can be alerted if any places we link to
are potentially down or very slow.