correct gender text values to be more accurate
This better refelects the situation with our gender data: currently
we only have `female` and `male` values (or none at all, where
gender is unknown). Although gender-balance.org allows for an
`other` value, we haven't yet had confirmed data using it, and
furthermore we can anticipate values where other values might
be used anyway. So stating anything other than the two current
values is aspirational but misleading.

But this change is necessary to (subtly) imply that we are
not _limiting_ our data to only use these two values, just
that currently those are the only values we have.